Move TestUtil.cs into OpenXmlPowerTools so that it can be used in interim development
