+++ OpenXmlPowerTools.Tests/OpenXmlPowerTools.Tests.csproj
@@ -71,7 +71,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RevisionAccepterTests.cs" />
     <Compile Include="SpreadsheetWriterTests.cs" />
-    <Compile Include="TestUtil.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
